backend/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.1.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
bleach==5.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.2
django-rest==0.8.7
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.39.4
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.1
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pyzmq==24.0.1
requests==2.31.0
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.1
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
